requirements.txt
appdirs==1.4.3
PyHyphen==3.0.1
six==1.12.0
